[Billing] throw some exceptions

+++ src/main/java/com/gem/support/service/impl/InvoiceServiceImpl.java
@@ -5,6 +5,7 @@ import com.gem.support.persistent.model.QInvoice;
 import com.gem.support.persistent.repository.InvoiceRepository;
 import com.gem.support.service.InvoiceService;
 import com.gem.support.service.dto.InvoiceDTO;
+import com.gem.support.service.exception.ResourceNotFoundException;
 import com.mysema.query.types.expr.BooleanExpression;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,6 +33,8 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     public void update(InvoiceDTO invoiceDTO) {
+        if(!invoiceRepository.exists(invoiceDTO.getId()))
+            throw new ResourceNotFoundException("requested invoice could not be found");
         Invoice invoice = new Invoice();
         BeanUtils.copyProperties(invoiceDTO, invoice);
         invoiceRepository.save(invoice);
@@ -39,12 +42,16 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     public void delete(String id) {
+        if(!invoiceRepository.exists(id))
+            throw new ResourceNotFoundException("requested invoice could not be found");
         invoiceRepository.delete(id);
     }
 
     @Override
     public InvoiceDTO findOne(String s) {
         Invoice invoice = invoiceRepository.findOne(s);
+        if(invoice == null)
+            throw new ResourceNotFoundException("requested invoice could not be found");
         InvoiceDTO dto = new InvoiceDTO();
         BeanUtils.copyProperties(invoice, dto);
         return dto;
